requirements.txt
openai>=1.30.0
groq>=0.9.0
requests>=2.31.0
google-generativeai>=0.7.0
cohere>=5.5.0
python-dotenv>=1.0.0
